Relocation refactoring
